ASI-289 add folio writer/reader, create ASpace marc_exporter
